lab5/requirements.txt
pymongo
cryptography
Flask
pyjwt
